moved SharedCode.c and type folder

+++ ptolemy/codegen/c/kernel/type/Array.c
@@ -57,7 +57,7 @@ Token Array_new(int size, int given, ...) {
             // convert the elements if needed.
             for (i = 0; i < given; i++) {
                 if (Array_get(result, i).type != elementType) {
-                    result.payload.Array->elements[i] = functionTable[(int)elementType][FUNC_convert](Array_get(result, i));
+                    result.payload.Array->elements[i] = functionTable[elementType][FUNC_convert](Array_get(result, i));
                 }
             }    
         }
@@ -76,7 +76,7 @@ Token Array_delete(Token token, ...) {
     // Delete each elements.
     for (i = 0; i < token.payload.Array->size; i++) {
     	element = Array_get(token, i);
-        functionTable[(int)element.type][FUNC_delete](element);
+        functionTable[element.type][FUNC_delete](element);
     }
     free(token.payload.Array->elements);
     free(token.payload.Array);
@@ -219,7 +219,7 @@ Token Array_multiply(Token this, ...) {
     result = Array_new(this.payload.Array->size, 0);
 	
     for (i = 0; i < this.payload.Array->size; i++) {
-        result.payload.Array->elements[i] = functionTable[(int)Array_get(this, i).type][FUNC_multiply](Array_get(this, i), Array_get(otherToken, i));
+        result.payload.Array->elements[i] = functionTable[Array_get(this, i).type][FUNC_multiply](Array_get(this, i), Array_get(otherToken, i));
     }
 
     va_end(argp);
@@ -338,7 +338,7 @@ Token Array_convert(Token token, ...) {
     for (i = 0; i < token.payload.Array->size; i++) {
         element = Array_get(token, i);
         if (targetType != element.type) {
-            result.payload.Array->elements[i] = functionTable[(int)targetType][FUNC_convert](element);
+            result.payload.Array->elements[i] = functionTable[targetType][FUNC_convert](element);
         } else {
             result.payload.Array->elements[i] = element;
         }
